Add check for deletion

+++ src/main/java/net/dv8tion/jda/internal/entities/ReceivedMessage.java
@@ -927,6 +927,8 @@ public class ReceivedMessage extends AbstractMessage
             else if (!sMember.hasPermission(gChan, Permission.MESSAGE_MANAGE))
                 throw new InsufficientPermissionException(gChan, Permission.MESSAGE_MANAGE);
         }
+        if (!type.canDelete())
+            throw new IllegalStateException("Cannot delete messages of type " + type);
         return channel.deleteMessageById(getIdLong());
     }
